+++ instagram_login_screen/lib/app_assets.dart
@@ -1,0 +1,4 @@
+class AppImg {
+  static const imgPath = 'assets/images';
+  static const imgLogo = '$imgPath/instagram_logo.png';
+}
