+++ instagram_login_screen/lib/screens/login/login_screen.dart
@@ -1,90 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:instagram_login_screen/app_assets.dart';
+import 'package:instagram_login_screen/commons/common_widgets.dart';
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({Key? key}) : super(key: key);
 
   @override
   _LoginScreenState createState() => _LoginScreenState();
-}
-
-// Text Field Template
-Widget buildTextField({required String text, bool hide = false}) {
-  return Container(
-    alignment: Alignment.centerLeft,
-    decoration: BoxDecoration(
-      color: Colors.white10,
-      borderRadius: BorderRadius.circular(10),
-    ),
-    height: 60,
-    child: Padding(
-      padding: const EdgeInsets.only(left: 15.0),
-      child: TextField(
-        obscureText: hide,
-        style: const TextStyle(
-          color: Colors.white60,
-          fontSize: 18,
-        ),
-        decoration: InputDecoration(
-          // filled: true,
-          // fillColor: Colors.transparent,
-          hintText: text,
-          hintStyle: const TextStyle(
-            color: Colors.white54,
-          ),
-          border: InputBorder.none,
-          // contentPadding: EdgeInsets.only(top: 14),
-        ),
-      ),
-    ),
-  );
-}
-
-Widget buildButton({required String text}) {
-  // TODO: full width
-
-  return ElevatedButton(
-    child: Text(
-      text,
-      style: const TextStyle(
-        color: Colors.white,
-      ),
-    ),
-    onPressed: () => print('$text Pressed'),
-    style: ElevatedButton.styleFrom(
-      primary: Colors.lightBlue,
-      padding: const EdgeInsets.symmetric(horizontal: 40, vertical: 20),
-      textStyle: const TextStyle(
-        color: Colors.white,
-        fontSize: 18,
-        fontWeight: FontWeight.w500,
-      ),
-    ),
-  );
-}
-
-Widget buildSignUpText() {
-  // TODO: add flexible - test on mobile
-  return Row(
-    mainAxisAlignment: MainAxisAlignment.center,
-    crossAxisAlignment: CrossAxisAlignment.end,
-    children: const <Widget>[
-      Text(
-        'Don\'t have an account?',
-        style: TextStyle(
-          color: Colors.white70,
-          fontSize: 16,
-        ),
-      ),
-      Text(
-        ' Sign Up',
-        style: TextStyle(
-          color: Colors.white70,
-          fontSize: 16,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-    ],
-  );
 }
 
 class _LoginScreenState extends State<LoginScreen> {
@@ -107,46 +29,53 @@ class _LoginScreenState extends State<LoginScreen> {
             ),
           ),
           child: Padding(
-            padding: const EdgeInsets.symmetric(
-              horizontal: 30,
-              vertical: 20,
+            padding: const EdgeInsets.only(
+              top: 78,
+              bottom: 10,
+              left: 30,
+              right: 30,
             ),
             child: Column(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
-                const Text(
-                  'Instagram',
-                  style: TextStyle(
-                    fontSize: 50,
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                // TODO: write common widget
-                buildTextField(text: 'Email'),
-                buildTextField(text: 'Password', hide: true),
-                buildButton(text: 'Log In'),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: const <Widget>[
-                    Text(
-                      'Forgot your password?',
-                      style: TextStyle(
-                        color: Colors.white70,
-                        fontSize: 16,
-                      ),
-                    ),
-                    Text(
-                      ' Get help signing in.',
-                      style: TextStyle(
-                        color: Colors.white70,
-                        fontSize: 16,
-                        fontWeight: FontWeight.bold,
-                      ),
+                Column(
+                  children: [
+                    Image.asset(AppImg.imgLogo),
+                    const SizedBox(height: 30),
+                    buildTextField(text: 'Email'),
+                    const SizedBox(height: 25),
+                    buildTextField(text: 'Password', hide: true),
+                    const SizedBox(height: 25),
+                    buildButton(text: 'Log In'),
+                    const SizedBox(height: 25),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: const <Widget>[
+                        Flexible(
+                          flex: 2,
+                          child: Text(
+                            'Forgot your password?',
+                            style: TextStyle(
+                              color: Colors.white70,
+                              fontSize: 16,
+                            ),
+                          ),
+                        ),
+                        Flexible(
+                          flex: 2,
+                          child: Text(
+                            ' Get help signing in.',
+                            style: TextStyle(
+                              color: Colors.white70,
+                              fontSize: 16,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ),
+                      ],
                     ),
                   ],
-                ),
-                // const Spacer(),
+                ), // const Spacer(),
                 buildSignUpText(),
               ],
             ),
